moved tests out of modules - Logging - Security
