+++ src/logging/tests.rs
@@ -8,9 +8,21 @@
 mod logging_tests {
     use super::super::*;
     use crate::logging::config::LoggingConfig;
+    use crate::logging::middleware::{LoggingMiddleware, ResponseFutureExt};
+    use crate::logging::structured::{LogFormat, LoggerConfig, RequestInfo, generate_trace_id};
     use crate::logging::test_logger;
+    use bytes::Bytes;
+    use http_body_util::Empty;
+    use hyper::{Method, Request, Response};
     use log::LevelFilter;
+    use std::collections::HashMap;
+    use std::future::Future;
+    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+    use std::pin::Pin;
     use std::sync::atomic::Ordering;
+    use std::task::{Context, Poll};
+    use std::time::Duration;
+    use std::time::{SystemTime, UNIX_EPOCH};
 
     // Helper function to reset logging state for testing
     fn reset_logging_state() {
@@ -565,5 +577,687 @@ mod logging_tests {
         // The returned error should be identical to the original
         assert_eq!(returned_error.code, 404);
         assert_eq!(returned_error.message, "Not found");
+    }
+
+    /* From `structured.rs` */
+    #[test]
+    fn test_logger_config_default() {
+        let config = LoggerConfig::default();
+
+        assert_eq!(config.format, LogFormat::Terminal);
+        assert_eq!(config.level, slog::Level::Info);
+        assert!(config.include_location);
+        assert!(config.include_thread_id);
+        assert!(config.static_fields.is_empty());
+    }
+
+    #[test]
+    fn test_logger_config_custom() {
+        let mut static_fields = HashMap::new();
+        static_fields.insert("service".to_string(), "foxy".to_string());
+        static_fields.insert("version".to_string(), "1.0.0".to_string());
+
+        let config = LoggerConfig {
+            format: LogFormat::Json,
+            level: slog::Level::Debug,
+            include_location: false,
+            include_thread_id: false,
+            static_fields,
+        };
+
+        assert_eq!(config.format, LogFormat::Json);
+        assert_eq!(config.level, slog::Level::Debug);
+        assert!(!config.include_location);
+        assert!(!config.include_thread_id);
+        assert_eq!(config.static_fields.len(), 2);
+        assert_eq!(
+            config.static_fields.get("service"),
+            Some(&"foxy".to_string())
+        );
+        assert_eq!(
+            config.static_fields.get("version"),
+            Some(&"1.0.0".to_string())
+        );
+    }
+
+    #[test]
+    fn test_logger_config_debug() {
+        let config = LoggerConfig::default();
+        let debug_str = format!("{config:?}");
+
+        assert!(debug_str.contains("LoggerConfig"));
+        assert!(debug_str.contains("format"));
+        assert!(debug_str.contains("level"));
+        assert!(debug_str.contains("include_location"));
+        assert!(debug_str.contains("include_thread_id"));
+        assert!(debug_str.contains("static_fields"));
+    }
+
+    #[test]
+    fn test_logger_config_clone() {
+        let mut config = LoggerConfig::default();
+        config
+            .static_fields
+            .insert("test".to_string(), "value".to_string());
+
+        let cloned = config.clone();
+
+        assert_eq!(config.format, cloned.format);
+        assert_eq!(config.level, cloned.level);
+        assert_eq!(config.include_location, cloned.include_location);
+        assert_eq!(config.include_thread_id, cloned.include_thread_id);
+        assert_eq!(config.static_fields, cloned.static_fields);
+    }
+
+    #[test]
+    fn test_log_format_equality() {
+        assert_eq!(LogFormat::Terminal, LogFormat::Terminal);
+        assert_eq!(LogFormat::Json, LogFormat::Json);
+        assert_ne!(LogFormat::Terminal, LogFormat::Json);
+        assert_ne!(LogFormat::Json, LogFormat::Terminal);
+    }
+
+    #[test]
+    fn test_log_format_debug() {
+        let terminal_debug = format!("{:?}", LogFormat::Terminal);
+        let json_debug = format!("{:?}", LogFormat::Json);
+
+        assert!(terminal_debug.contains("Terminal"));
+        assert!(json_debug.contains("Json"));
+    }
+
+    #[test]
+    fn test_log_format_clone() {
+        let terminal = LogFormat::Terminal;
+        let json = LogFormat::Json;
+
+        let terminal_clone = terminal.clone();
+        let json_clone = json.clone();
+
+        assert_eq!(terminal, terminal_clone);
+        assert_eq!(json, json_clone);
+    }
+
+    #[test]
+    fn test_request_info_creation() {
+        let start_time = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_millis();
+
+        let request_info = RequestInfo {
+            trace_id: "test-trace-id".to_string(),
+            method: "GET".to_string(),
+            path: "/api/test".to_string(),
+            remote_addr: "127.0.0.1:8080".to_string(),
+            user_agent: "test-agent".to_string(),
+            start_time_ms: start_time,
+        };
+
+        assert_eq!(request_info.trace_id, "test-trace-id");
+        assert_eq!(request_info.method, "GET");
+        assert_eq!(request_info.path, "/api/test");
+        assert_eq!(request_info.remote_addr, "127.0.0.1:8080");
+        assert_eq!(request_info.user_agent, "test-agent");
+        assert_eq!(request_info.start_time_ms, start_time);
+    }
+
+    #[test]
+    fn test_request_info_elapsed_ms() {
+        let start_time = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_millis()
+            .saturating_sub(100); // 100ms ago
+
+        let request_info = RequestInfo {
+            trace_id: "test".to_string(),
+            method: "GET".to_string(),
+            path: "/test".to_string(),
+            remote_addr: "127.0.0.1".to_string(),
+            user_agent: "test".to_string(),
+            start_time_ms: start_time,
+        };
+
+        let elapsed = request_info.elapsed_ms();
+        // Should be approximately 100ms, but allow for some variance
+        assert!((90..=200).contains(&elapsed));
+    }
+
+    #[test]
+    fn test_request_info_elapsed_ms_future_time() {
+        // Test with a future start time (should return 0 due to saturating_sub)
+        let future_time = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_millis()
+            .saturating_add(1000); // 1 second in the future
+
+        let request_info = RequestInfo {
+            trace_id: "test".to_string(),
+            method: "GET".to_string(),
+            path: "/test".to_string(),
+            remote_addr: "127.0.0.1".to_string(),
+            user_agent: "test".to_string(),
+            start_time_ms: future_time,
+        };
+
+        let elapsed = request_info.elapsed_ms();
+        assert_eq!(elapsed, 0);
+    }
+
+    #[test]
+    fn test_request_info_debug() {
+        let request_info = RequestInfo {
+            trace_id: "debug-test".to_string(),
+            method: "POST".to_string(),
+            path: "/debug".to_string(),
+            remote_addr: "192.168.1.1".to_string(),
+            user_agent: "debug-agent".to_string(),
+            start_time_ms: 1_234_567_890,
+        };
+
+        let debug_str = format!("{request_info:?}");
+        assert!(debug_str.contains("RequestInfo"));
+        assert!(debug_str.contains("debug-test"));
+        assert!(debug_str.contains("POST"));
+        assert!(debug_str.contains("/debug"));
+        assert!(debug_str.contains("192.168.1.1"));
+        assert!(debug_str.contains("debug-agent"));
+        assert!(debug_str.contains("1234567890"));
+    }
+
+    #[test]
+    fn test_request_info_clone() {
+        let original = RequestInfo {
+            trace_id: "clone-test".to_string(),
+            method: "PUT".to_string(),
+            path: "/clone".to_string(),
+            remote_addr: "10.0.0.1".to_string(),
+            user_agent: "clone-agent".to_string(),
+            start_time_ms: 9_876_543_210,
+        };
+
+        let cloned = original.clone();
+
+        assert_eq!(original.trace_id, cloned.trace_id);
+        assert_eq!(original.method, cloned.method);
+        assert_eq!(original.path, cloned.path);
+        assert_eq!(original.remote_addr, cloned.remote_addr);
+        assert_eq!(original.user_agent, cloned.user_agent);
+        assert_eq!(original.start_time_ms, cloned.start_time_ms);
+    }
+
+    #[test]
+    fn test_generate_trace_id() {
+        let trace_id1 = generate_trace_id();
+        let trace_id2 = generate_trace_id();
+
+        // Should be valid UUIDs
+        assert!(uuid::Uuid::parse_str(&trace_id1).is_ok());
+        assert!(uuid::Uuid::parse_str(&trace_id2).is_ok());
+
+        // Should be different
+        assert_ne!(trace_id1, trace_id2);
+
+        // Should be the right length (UUID v4 string format)
+        assert_eq!(trace_id1.len(), 36);
+        assert_eq!(trace_id2.len(), 36);
+    }
+
+    #[test]
+    fn test_generate_trace_id_format() {
+        let trace_id = generate_trace_id();
+
+        // Should match UUID format: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
+        let parts: Vec<&str> = trace_id.split('-').collect();
+        assert_eq!(parts.len(), 5);
+        assert_eq!(parts[0].len(), 8);
+        assert_eq!(parts[1].len(), 4);
+        assert_eq!(parts[2].len(), 4);
+        assert_eq!(parts[3].len(), 4);
+        assert_eq!(parts[4].len(), 12);
+
+        // All parts should be hexadecimal
+        for part in parts {
+            assert!(part.chars().all(|c| c.is_ascii_hexdigit() || c == '-'));
+        }
+    }
+
+    #[test]
+    fn test_init_global_logger_terminal() {
+        let config = LoggerConfig {
+            format: LogFormat::Terminal,
+            level: slog::Level::Info,
+            include_location: true,
+            include_thread_id: true,
+            static_fields: HashMap::new(),
+        };
+
+        // Test that the function returns a guard without panicking
+        let _guard = init_global_logger(&config);
+
+        // Test that the guard exists (this validates the logger was created)
+        // We don't test actual logging to avoid global state conflicts in parallel tests
+    }
+
+    #[test]
+    fn test_init_global_logger_json() {
+        let config = LoggerConfig {
+            format: LogFormat::Json,
+            level: slog::Level::Debug,
+            include_location: false,
+            include_thread_id: false,
+            static_fields: HashMap::new(),
+        };
+
+        // Test that the function returns a guard without panicking
+        let _guard = init_global_logger(&config);
+
+        // Test that the guard exists (this validates the logger was created)
+        // We don't test actual logging to avoid global state conflicts in parallel tests
+    }
+
+    #[test]
+    fn test_init_global_logger_with_static_fields() {
+        let mut static_fields = HashMap::new();
+        static_fields.insert("service".to_string(), "test-service".to_string());
+        static_fields.insert("environment".to_string(), "test".to_string());
+
+        let config = LoggerConfig {
+            format: LogFormat::Terminal,
+            level: slog::Level::Warning,
+            include_location: true,
+            include_thread_id: true,
+            static_fields,
+        };
+
+        // Test that the function returns a guard without panicking
+        let _guard = init_global_logger(&config);
+
+        // Test that the guard exists (this validates the logger was created)
+        // We don't test actual logging to avoid global state conflicts in parallel tests
+    }
+
+    /* From `middleware.rs` */
+    fn create_test_config() -> LoggingConfig {
+        LoggingConfig {
+            structured: false,
+            format: "terminal".to_string(),
+            level: "info".to_string(),
+            include_location: true,
+            include_thread_id: true,
+            include_trace_id: true,
+            propagate_trace_id: false,
+            trace_id_header: "x-trace-id".to_string(),
+            static_fields: std::collections::HashMap::new(),
+        }
+    }
+
+    fn create_test_config_structured() -> LoggingConfig {
+        LoggingConfig {
+            structured: true,
+            format: "json".to_string(),
+            level: "info".to_string(),
+            include_location: true,
+            include_thread_id: true,
+            include_trace_id: true,
+            propagate_trace_id: false,
+            trace_id_header: "x-trace-id".to_string(),
+            static_fields: std::collections::HashMap::new(),
+        }
+    }
+
+    fn create_test_config_with_propagation() -> LoggingConfig {
+        LoggingConfig {
+            structured: false,
+            format: "terminal".to_string(),
+            level: "info".to_string(),
+            include_location: true,
+            include_thread_id: true,
+            include_trace_id: true,
+            propagate_trace_id: true,
+            trace_id_header: "x-trace-id".to_string(),
+            static_fields: std::collections::HashMap::new(),
+        }
+    }
+
+    fn create_test_request() -> Request<Empty<Bytes>> {
+        Request::builder()
+            .method(Method::GET)
+            .uri("/test/path")
+            .header("user-agent", "test-agent/1.0")
+            .body(Empty::<Bytes>::new())
+            .unwrap()
+    }
+
+    fn create_test_request_with_trace_id(trace_id: &str) -> Request<Empty<Bytes>> {
+        Request::builder()
+            .method(Method::GET)
+            .uri("/test/path")
+            .header("user-agent", "test-agent/1.0")
+            .header("x-trace-id", trace_id)
+            .body(Empty::<Bytes>::new())
+            .unwrap()
+    }
+
+    fn create_test_socket_addr() -> SocketAddr {
+        SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)), 8080)
+    }
+
+    #[tokio::test]
+    async fn test_logging_middleware_new() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+
+        // Test that the middleware is created successfully
+        assert!(!middleware.config().structured); // Use getter
+        assert!(!middleware.config().propagate_trace_id); // Use getter
+    }
+
+    #[tokio::test]
+    async fn test_process_request_basic() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+        let request = create_test_request();
+        let remote_addr = Some(create_test_socket_addr());
+
+        let (processed_req, request_info) = middleware.process(request, remote_addr);
+
+        // Verify request is returned unchanged
+        assert_eq!(processed_req.method(), Method::GET);
+        assert_eq!(processed_req.uri().path(), "/test/path");
+
+        // Verify request info is populated
+        assert_eq!(request_info.method, "GET");
+        assert_eq!(request_info.path, "/test/path");
+        assert_eq!(request_info.remote_addr, "192.168.1.100:8080");
+        assert_eq!(request_info.user_agent, "test-agent/1.0");
+        assert!(!request_info.trace_id.is_empty());
+    }
+
+    #[tokio::test]
+    async fn test_process_request_no_remote_addr() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+        let request = create_test_request();
+
+        let (_, request_info) = middleware.process(request, None);
+
+        assert_eq!(request_info.remote_addr, "unknown");
+    }
+
+    #[tokio::test]
+    async fn test_process_request_no_user_agent() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+        let request = Request::builder()
+            .method(Method::POST)
+            .uri("/api/test")
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let (_, request_info) = middleware.process(request, None);
+
+        assert_eq!(request_info.method, "POST");
+        assert_eq!(request_info.path, "/api/test");
+        assert_eq!(request_info.user_agent, "unknown");
+    }
+
+    #[tokio::test]
+    async fn test_process_request_with_trace_propagation() {
+        let config = create_test_config_with_propagation();
+        let middleware = LoggingMiddleware::new(config);
+        let existing_trace_id = "existing-trace-123";
+        let request = create_test_request_with_trace_id(existing_trace_id);
+
+        let (_, request_info) = middleware.process(request, None);
+
+        assert_eq!(request_info.trace_id, existing_trace_id);
+    }
+
+    #[tokio::test]
+    async fn test_process_request_without_trace_propagation() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+        let request = create_test_request_with_trace_id("existing-trace-123");
+
+        let (_, request_info) = middleware.process(request, None);
+
+        // Should generate new trace ID, not use existing one
+        assert_ne!(request_info.trace_id, "existing-trace-123");
+        assert!(!request_info.trace_id.is_empty());
+    }
+
+    #[tokio::test]
+    async fn test_process_request_invalid_trace_header() {
+        let config = create_test_config_with_propagation();
+        let middleware = LoggingMiddleware::new(config);
+
+        // Create a request with an invalid trace header value that can't be parsed as UTF-8
+        // We'll use a valid header construction but with an empty value to test the fallback
+        let request = Request::builder()
+            .method(Method::GET)
+            .uri("/test")
+            .header("x-trace-id", "") // Empty trace ID should trigger fallback
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let (_, request_info) = middleware.process(request, None);
+
+        // Should generate new trace ID when existing one is empty/invalid
+        assert!(!request_info.trace_id.is_empty());
+        assert_ne!(request_info.trace_id, "");
+    }
+
+    #[test]
+    fn test_log_response_basic() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+
+        let response = Response::builder()
+            .status(200)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let request_info = RequestInfo {
+            trace_id: "test-trace-123".to_string(),
+            method: "GET".to_string(),
+            path: "/test".to_string(),
+            remote_addr: "192.168.1.1".to_string(),
+            user_agent: "test-agent".to_string(),
+            start_time_ms: std::time::SystemTime::now()
+                .duration_since(std::time::UNIX_EPOCH)
+                .unwrap()
+                .as_millis(),
+        };
+
+        // This should not panic
+        middleware.log_response(&response, &request_info, Some(Duration::from_millis(50)));
+    }
+
+    #[test]
+    fn test_log_response_structured() {
+        let config = create_test_config_structured();
+        let middleware = LoggingMiddleware::new(config);
+
+        let response = Response::builder()
+            .status(404)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let request_info = RequestInfo {
+            trace_id: "test-trace-456".to_string(),
+            method: "POST".to_string(),
+            path: "/api/users".to_string(),
+            remote_addr: "10.0.0.1".to_string(),
+            user_agent: "curl/7.68.0".to_string(),
+            start_time_ms: std::time::SystemTime::now()
+                .duration_since(std::time::UNIX_EPOCH)
+                .unwrap()
+                .as_millis(),
+        };
+
+        // This should not panic
+        middleware.log_response(&response, &request_info, None);
+    }
+
+    #[test]
+    fn test_log_response_no_upstream_duration() {
+        let config = create_test_config();
+        let middleware = LoggingMiddleware::new(config);
+
+        let response = Response::builder()
+            .status(500)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let request_info = RequestInfo {
+            trace_id: "test-trace-789".to_string(),
+            method: "DELETE".to_string(),
+            path: "/api/resource/123".to_string(),
+            remote_addr: "172.16.0.1".to_string(),
+            user_agent: "Mozilla/5.0".to_string(),
+            start_time_ms: std::time::SystemTime::now()
+                .duration_since(std::time::UNIX_EPOCH)
+                .unwrap()
+                .as_millis(),
+        };
+
+        // This should not panic and handle None upstream duration
+        middleware.log_response(&response, &request_info, None);
+    }
+
+    // Mock future for testing TracedResponseFuture
+    struct MockResponseFuture {
+        response: Option<Result<Response<Empty<Bytes>>, &'static str>>,
+    }
+
+    impl MockResponseFuture {
+        fn new_ok(response: Response<Empty<Bytes>>) -> Self {
+            Self {
+                response: Some(Ok(response)),
+            }
+        }
+
+        fn new_err(error: &'static str) -> Self {
+            Self {
+                response: Some(Err(error)),
+            }
+        }
+    }
+
+    impl Future for MockResponseFuture {
+        type Output = Result<Response<Empty<Bytes>>, &'static str>;
+
+        fn poll(mut self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Self::Output> {
+            Poll::Ready(self.response.take().unwrap())
+        }
+    }
+
+    impl Unpin for MockResponseFuture {}
+
+    #[tokio::test]
+    async fn test_traced_response_future_success_with_trace_id() {
+        let response = Response::builder()
+            .status(200)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let future = MockResponseFuture::new_ok(response);
+        let traced_future =
+            future.with_trace_id("test-trace-123".to_string(), "x-trace-id".to_string(), true);
+
+        let result = traced_future.await;
+        assert!(result.is_ok());
+
+        let response = result.unwrap();
+        assert_eq!(response.status(), 200);
+        assert!(response.headers().contains_key("x-trace-id"));
+        assert_eq!(
+            response.headers().get("x-trace-id").unwrap(),
+            "test-trace-123"
+        );
+    }
+
+    #[tokio::test]
+    async fn test_traced_response_future_success_without_trace_id() {
+        let response = Response::builder()
+            .status(201)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let future = MockResponseFuture::new_ok(response);
+        let traced_future = future.with_trace_id(
+            "test-trace-456".to_string(),
+            "x-trace-id".to_string(),
+            false, // Don't include trace ID
+        );
+
+        let result = traced_future.await;
+        assert!(result.is_ok());
+
+        let response = result.unwrap();
+        assert_eq!(response.status(), 201);
+        assert!(!response.headers().contains_key("x-trace-id"));
+    }
+
+    #[tokio::test]
+    async fn test_traced_response_future_error() {
+        let future = MockResponseFuture::new_err("test error");
+        let traced_future =
+            future.with_trace_id("test-trace-789".to_string(), "x-trace-id".to_string(), true);
+
+        let result = traced_future.await;
+        assert!(result.is_err());
+        assert_eq!(result.unwrap_err(), "test error");
+    }
+
+    #[tokio::test]
+    async fn test_traced_response_future_invalid_header_name() {
+        let response = Response::builder()
+            .status(200)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let future = MockResponseFuture::new_ok(response);
+        let traced_future = future.with_trace_id(
+            "test-trace-123".to_string(),
+            "invalid header name with spaces".to_string(), // Invalid header name
+            true,
+        );
+
+        let result = traced_future.await;
+        assert!(result.is_ok());
+
+        let response = result.unwrap();
+        // Should fallback to x-trace-id header
+        assert!(response.headers().contains_key("x-trace-id"));
+    }
+
+    #[tokio::test]
+    async fn test_traced_response_future_invalid_header_value() {
+        let response = Response::builder()
+            .status(200)
+            .body(Empty::<Bytes>::new())
+            .unwrap();
+
+        let future = MockResponseFuture::new_ok(response);
+        let traced_future = future.with_trace_id(
+            "\x00\x01\x02".to_string(), // Invalid header value
+            "x-trace-id".to_string(),
+            true,
+        );
+
+        let result = traced_future.await;
+        assert!(result.is_ok());
+
+        let response = result.unwrap();
+        assert!(response.headers().contains_key("x-trace-id"));
+        // Should fallback to "invalid-trace-id"
+        assert_eq!(
+            response.headers().get("x-trace-id").unwrap(),
+            "invalid-trace-id"
+        );
     }
 }
